requirements.txt
alembic==1.14.0
aniso8601==9.0.1
attrs==24.2.0
azure-core==1.32.0
azure-storage-blob==12.23.1
azure-storage-file-share==12.19.0
blinker==1.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dnspython==2.7.0
email-validator==2.0.0
Flask==2.2.5
Flask-Dance==7.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Minify==0.37
flask-restx==1.0.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.10
infinity==1.5
intervals==0.9.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.2
jsmin==3.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lesscpy==0.15.1
Mako==1.3.6
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
ply==3.11
pycparser==2.22
PyJWT==2.4.0
pyodbc==5.2.0
python-dotenv==0.19.2
pytz==2024.2
rcssmin==1.1.3
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
setuptools==75.4.0
six==1.16.0
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
URLObject==2.4.3
validators==0.34.0
Werkzeug==2.3.7
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
xxhash==3.5.0
